adapt list-expense-items directive to a accept a readonly parameter

+++ src/js/expense-items/list-expense-items.directive.js
@@ -9,7 +9,8 @@ function($modal, $filter, $timeout, spinnerService, globalMessagesService, expen
 		templateUrl: 'expense-items/list-expense-items.directive.tpl.html',
 		scope: {
 			expenseUid: '=',
-			expenseItems: '='
+			expenseItems: '=',
+			readonly: '='
 		},
 		link: function($scope) {
 
@@ -23,7 +24,11 @@ function($modal, $filter, $timeout, spinnerService, globalMessagesService, expen
 					for(var i=0; i<response.data.length; i++) {
 						// delete all expense-items with state initial (not finished creation)
 						if(response.data[i].state === 'INITIAL') {
-							expenseItemsRestService.deleteExpenseItem(response.data[i].uid);
+							// if an admin or finance-admin looks at a rejected expense, which
+							// is again in edit mode, initial expenses should not be deleted
+							if(!$scope.readonly) {
+								expenseItemsRestService.deleteExpenseItem(response.data[i].uid);
+							}
 						}
 						// show all others
 						else {
@@ -40,61 +45,77 @@ function($modal, $filter, $timeout, spinnerService, globalMessagesService, expen
 				updateTable();
 			});
 
-			$scope.editExpenseItem = function(expenseItemUid) {
-				var modalInstance = $modal.open({
-					templateUrl: 'expense-items/edit-expense-item.tpl.html',
-					controller: 'EditExpenseItemController',
-					resolve: {
-						expenseItemUid: function() {
-							return expenseItemUid;
-						}
-					},
-					// prevent closing by accident:
-					backdrop: 'static',
-					keyboard: false
-				});
+			if(!$scope.readonly) {
 
-				modalInstance.result.then()['finally'](updateTable);
-			};
-
-			$scope.deleteExpenseItem = function(expenseItemUid) {
-				globalMessagesService.confirmWarning("reimbursement.expense-item.deleteConfirmTitle",
-					"reimbursement.expense-item.deleteConfirmMessage").then(function() {
-
-					expenseItemsRestService.deleteExpenseItem(expenseItemUid).then(updateTable);
-				});
-			};
-
-			$scope.addExpenseItem = function() {
-				expenseItemsRestService.getCostCategories().then(function(response) {
-					var preSelectedCategoryUid = response.data[0].uid;
-
-					expenseItemsRestService.postExpenseItem($scope.expenseUid, {
-						date: $filter('date')(new Date(), 'yyyy-MM-dd'),
-						costCategoryUid: preSelectedCategoryUid,
-						currency: 'CHF',
-					}).then(function(response) {
-
-						var modalInstance = $modal.open({
-							templateUrl: 'expense-items/add-expense-item.tpl.html',
-							controller: 'AddExpenseItemController',
-							resolve: {
-								expenseItemUid: function() {
-									return response.data.uid;
-								}
-							},
-							// prevent closing by accident:
-							backdrop: 'static',
-							keyboard: false
-						});
-
-						modalInstance.result.then()['finally'](updateTable);
-
+				$scope.editExpenseItem = function(expenseItemUid) {
+					var modalInstance = $modal.open({
+						templateUrl: 'expense-items/edit-expense-item.tpl.html',
+						controller: 'EditExpenseItemController',
+						resolve: {
+							expenseItemUid: function() {
+								return expenseItemUid;
+							}
+						},
+						// prevent closing by accident:
+						backdrop: 'static',
+						keyboard: false
 					});
 
-				});
-			};
+					modalInstance.result.then()['finally'](updateTable);
+				};
 
+				$scope.deleteExpenseItem = function(expenseItemUid) {
+					globalMessagesService.confirmWarning("reimbursement.expense-item.deleteConfirmTitle",
+						"reimbursement.expense-item.deleteConfirmMessage").then(function() {
+
+						expenseItemsRestService.deleteExpenseItem(expenseItemUid).then(updateTable);
+					});
+				};
+
+				$scope.addExpenseItem = function() {
+					expenseItemsRestService.getCostCategories().then(function(response) {
+						var preSelectedCategoryUid = response.data[0].uid;
+
+						expenseItemsRestService.postExpenseItem($scope.expenseUid, {
+							date: $filter('date')(new Date(), 'yyyy-MM-dd'),
+							costCategoryUid: preSelectedCategoryUid,
+							currency: 'CHF',
+						}).then(function(response) {
+
+							var modalInstance = $modal.open({
+								templateUrl: 'expense-items/add-expense-item.tpl.html',
+								controller: 'AddExpenseItemController',
+								resolve: {
+									expenseItemUid: function() {
+										return response.data.uid;
+									}
+								},
+								// prevent closing by accident:
+								backdrop: 'static',
+								keyboard: false
+							});
+
+							modalInstance.result.then()['finally'](updateTable);
+
+						});
+
+					});
+				};
+			}
+
+			else {
+				$scope.viewExpenseItem = function(expenseItemUid) {
+					$modal.open({
+						templateUrl: 'expense-items/view-expense-item.tpl.html',
+						controller: 'ViewExpenseItemController',
+						resolve: {
+							expenseItemUid: function() {
+								return expenseItemUid;
+							}
+						}
+					});
+				};
+			}
 		}
 	};
